templates/fastapi/pyproject.toml
[tool.poetry]
name = "pyadmin"
version = "0.1.0"
description = ""
authors = ["wwqdrh <huiloademail@163.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
tortoise-orm = {extras = ["aiomysql"], version = "^0.17.4"}
Jinja2 = "^3.0.1"
unicorn = "^1.0.3"
aerich = "^0.5.3"
aioredis = "^1.3.1"
jose = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
